♪ Build-scripts and a Rakefile
Now you can compile a font! Works with both FontForge and the Adobe Font
Development Kit for OpenType. The former is what we aim for as our main
compiler, as the latter is closed-source, but to have both is great for
testing the UFO spec.

Run ''rake'' to generate the font. ''rake diagnostics'' will give you an
overview of your current build environment and advise you how to proceed if it
can’t find the right build tools.

If you want more control you can use tools/ufo2otf.py which provides a command
line interface: you can choose input and output files, and which compiler to
use.