'rake package' produces a release package as a zip archive
A build folder is created to put all the assets into before zipping
A white-listing approach is being taken; i.e. files need to be
specifically moved into the build folder before they get zipped.

Both the UFO sources and the generated OTF get marked with the version
number as based on the git

Note there is a pesky bug in this Rakefile:
We use a zip command like

zip -r OpenBaskerville-0.0.56.zip build/OpenBaskerville-0.56

Which produces a zipfile which includes the build folder.
How would one go about to get just the OpenBaskerville-0.0.56
folder in there?

https://github.com/klepas/open-baskerville/issues/10