Merging in glyphs from Century Catalogue
This Morris Fuller Benton design was digitised by Raph Levien,
http://levien.com/type/myfonts/
who allowed us to use them for Open Baskerville.
http://openbaskerville.lighthouseapp.com/projects/24633/messages/15022-greetings-from-raph

I scaled the glyphs by 90% which seems to be more or less correct,
then rounded the points to integers.

This is not mean to be correct out of the box, tweaking is required—
but I figure it good to commit already so we can tweak together.

The curly braces are too low, the lesser than greater then
signs too thin… The square brackets look about right.

The dashes should be made to match the hyphen.

The ordinal, degree, currency look about right,
as do other geometrical glyphs.

What do you think of the single and double quote?

The divide glyph is unsymmetric and needs to be fixed.

Paragraph and section might need slight restyling
to fit into Open Baskerville style.